+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.0
 ---
 
+ * Change the default value of the `$requireTld` option of the `Url` constraint to `true`
  * Add method `getGroupProvider()` to `ClassMetadataInterface`
  * Replace `__sleep/wakeup()` by `__(un)serialize()`  on `GenericMetadata` implementations
  * Remove the `getRequiredOptions()` and `getDefaultOption()` methods from the `All`, `AtLeastOneOf`, `CardScheme`, `Collection`,